some errors fixed in product and order services

+++ services/productService/src/controller/productController.ts
@@ -6,6 +6,9 @@ import Product from "../model/productModel"
 const getProducts = async(req:Request,res:Response)=>{
     try {
         const products = await Product.find()
+        if(products.length===0){
+            return res.json({message:"products is empty"})
+        }
         return res.status(200).json({products})
     } catch (error) {
         console.error('Error founded in product getting side ',error);
@@ -21,7 +24,7 @@ const productAdd = async(req:Request,res:Response)=>{
         if(existingProduct){
             return res.json({message:"this product is already existed"})
         }
-        const newProduct = Product.create({
+        const newProduct = await Product.create({
             name,
             category,
             quantity,
@@ -40,8 +43,11 @@ const productAdd = async(req:Request,res:Response)=>{
 //       product edit functionality
 const productEdit = async(req:Request,res:Response)=>{
     const {name,category,price,quantity}:{name:string;category:string;price:number;quantity:number} = req.body
-    const productId = req.params._id
+    const productId = req.params.id;
     console.log(productId,'productId')
+    if(!productId){
+        return res.json({message:"not getting productId"})
+    }
     try {
         const existingProduct = await Product.findOne({name:name,_id:{$ne:productId}})
         if(existingProduct){
@@ -63,10 +69,19 @@ const productEdit = async(req:Request,res:Response)=>{
 
 //         product delete functionality
 const productDelete = async(req:Request,res:Response)=>{
-    const productId = req.params._id
+    const productId = req.params.id
+    console.log(productId,'product id ');
+    if(!productId){
+        return res.json({message:"product id is not defined"})
+    }
     try {
+        const aa = await Product.findById(productId)
+        console.log(aa,'***    aaa      ***');
+        if(!aa){
+            return res.json({message:"product is not available "})
+        }
         await Product.findByIdAndDelete(productId)
-        res.status(200).json({message:"product delete successfully"})
+        return res.status(200).json({message:"product delete successfully"})
     } catch (error) {
         console.error('Error deleting product',error);
